requirements.txt
asgiref==3.2.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-model-utils==4.0.0
django-mysql==3.8.1
django-notifications-hq==1.6.0
djangorestframework==3.11.0
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
jsonfield==3.1.0
mysqlclient==1.4.6
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.10.1
pytz==2019.3
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.0
swapper==1.1.2.post1
urllib3==1.25.9
